+++ releases.md
@@ -1,6 +1,6 @@
 # Releases
 
-## Unreleased
+## v0.35.2
 
   - Tidy up implementation of `read_line?` to handle line length errors and protocol violations more clearly.
   - Improve error handling for unexpected connection closures (`Errno::ECONNRESET`) in `read_line?`.